Overall took care of the warning, buff Bloody shoot, revert the roguelike data, buff copper armor, implement Enchantment ( rn it do nothing ) fix a problem where playing another RPG mod cause perk UI to not appear, update desc and build num, seal most stuff up in ItemPlayerHandle, comment add recipe in enraged spawner, card no longer drop from any monster, Orb of alchemist get buff and Jungle Armor now have it's set bonus

+++ Common/Utils/BossRushUtils.Vector2.cs
@@ -55,14 +55,14 @@ namespace BossRush {
 			X = Math.Abs(newVelocity.X);
 			Y = Math.Abs(newVelocity.Y);
 		}
-		public static Vector2 Vector2DistributeEvenly(this Vector2 vec, float ProjectileAmount, float rotation, int i) {
+		public static Vector2 Vector2DistributeEvenly(this Vector2 vec, float ProjectileAmount, float rotation, float i) {
 			if (ProjectileAmount > 1) {
 				rotation = MathHelper.ToRadians(rotation);
 				return vec.RotatedBy(MathHelper.Lerp(rotation * .5f, rotation * -.5f, i / ProjectileAmount));
 			}
 			return vec;
 		}
-		public static Vector2 Vector2DistributeEvenlyPlus(this Vector2 vec, float ProjectileAmount, float rotation, int i) {
+		public static Vector2 Vector2DistributeEvenlyPlus(this Vector2 vec, float ProjectileAmount, float rotation, float i) {
 			if (ProjectileAmount > 1) {
 				rotation = MathHelper.ToRadians(rotation);
 				return vec.RotatedBy(MathHelper.Lerp(rotation * .5f, rotation * -.5f, i / (ProjectileAmount - 1f)));
